requirements.txt
Django~=3.1.6
pip~=21.0.1
django-ckeditor~=6.0.0
django-tinymce
psycopg2